requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Configuration
python-dotenv==1.0.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Logging (built-in, but may need structured logging)
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Development
ipython==8.18.1
